Update Firestore IFF the user is authenticated

+++ client/src/store/networks/actions/deletePerson.ts
@@ -19,41 +19,49 @@ import { setNetworkLoading } from "./setNetworkLoading"
  */
 
 export const deletePerson = (networkId: string, personId: string): AppThunk => {
-  return async (dispatch) => {
+  return async (dispatch, getState) => {
     dispatch(setNetworkLoading(true))
 
     try {
-      /* Remove the Person from the Network */
-      const networkDoc = networksCollection.doc(networkId)
-      const networkData = (await networkDoc.get()).data() as INetwork
-      const peopleWithoutDeletedPerson = networkData.personIds.filter(
-        (id) => id !== personId,
-      )
+      /* Update the database if the user is authenticated */
+      const uid = getState().auth.userId
+      if (uid) {
+        /* Ensure the network exists */
+        const networkDoc = await networksCollection.doc(networkId).get()
+        if (!networkDoc.exists)
+          throw new Error(`Network ${networkId} does not exist`)
 
-      /* Update the Network's personIds field with the new list of person IDs */
-      await networkDoc.update({ personIds: peopleWithoutDeletedPerson })
+        /* Remove the Person from the Network */
+        const networkData = networkDoc.data() as INetwork
+        const peopleWithoutDeletedPerson = networkData.personIds.filter(
+          (id) => id !== personId,
+        )
 
-      /* Get the Person's document*/
-      const personDoc = peopleCollection.doc(personId)
-      const personData = (await (await personDoc.get()).data()) as IPerson
+        /* Update the Network's personIds field with the new list of person IDs */
+        await networkDoc.ref.update({ personIds: peopleWithoutDeletedPerson })
 
-      /* Remove all relationships including the deleted Person */
-      const relationshipUpdates: Promise<void>[] = Object.keys(
-        personData.relationships,
-      ).map(async (relationshipId) => {
-        const otherPersonDoc = peopleCollection.doc(relationshipId)
-        const otherPersonData = (await otherPersonDoc.get()).data() as IPerson
-        const updatedRelationship: IRelationships = {
-          ...otherPersonData.relationships,
-        }
-        delete updatedRelationship[personId]
-        return otherPersonDoc.update({ relationships: updatedRelationship })
-      })
+        /* Get the Person's document*/
+        const personDoc = peopleCollection.doc(personId)
+        const personData = (await (await personDoc.get()).data()) as IPerson
 
-      await Promise.all(relationshipUpdates)
+        /* Remove all relationships including the deleted Person */
+        const relationshipUpdates: Promise<void>[] = Object.keys(
+          personData.relationships,
+        ).map(async (relationshipId) => {
+          const otherPersonDoc = peopleCollection.doc(relationshipId)
+          const otherPersonData = (await otherPersonDoc.get()).data() as IPerson
+          const updatedRelationship: IRelationships = {
+            ...otherPersonData.relationships,
+          }
+          delete updatedRelationship[personId]
+          return otherPersonDoc.update({ relationships: updatedRelationship })
+        })
 
-      /* Delete the Person's document */
-      await personDoc.delete()
+        await Promise.all(relationshipUpdates)
+
+        /* Delete the Person's document */
+        await personDoc.delete()
+      }
 
       /* Update state accordingly with networkId and personId */
       const action: IDeletePersonByIdAction = {
